FastGradientProjection

Documentation for FastGradientProjection.

FastGradientProjection.FGP

FastGradientProjection.FGP!

FastGradientProjection.GP

FastGradientProjection.GP!

FastGradientProjection.FGP! — Method

FGP!(b, λ, N; lower_bound=-Inf, upper_bound=Inf, TV="iso")

Same as FGP, but operates in-place on b.

FastGradientProjection.FGP — Method

FGP(b, λ, N; lower_bound=-Inf, upper_bound=Inf, TV="iso")

return denoised (volume) image b based on the minimization problem $\min_{\mathbf{x}\in{C}}\|\mathbf{x} - \mathbf{b}\|^{2}_{F} + 2\lambda\operatorname{TV}(\mathbf{x}).$

Arguments

b: input (volume) image.

λ: regularization parameter.

N: number of iterations.

lower_bound=-Inf: upper bound of the convex closed set $C$. If b is a complex array, projection to $C$ is not performed.

upper_bound=Inf: lower bound of the convex closed set $C$.

TV="iso": if TV="iso" (default), denoising is based on isotropic TV. To specify anisotropic TV, set TV="aniso".

FastGradientProjection.GP! — Method

GP!(b, λ, N; lower_bound=-Inf, upper_bound=Inf, TV="iso")

Same as GP, but operates in-place on b.

FastGradientProjection.GP — Method

GP(b, λ, N; lower_bound=-Inf, upper_bound=Inf, TV="iso")

return denoised (volume) image b. Compared to FGP, convergence is slower, but memory usage is lower.